Use gint64 to represent time in MonoStopwatch to prevent overflows.
Commit migrated from https://github.com/mono/mono/commit/ed541952ef277ffe10416f67dea1e8c718ba787c

+++ src/mono/mono/utils/mono-time.h
@@ -15,7 +15,7 @@ gint64  mono_100ns_datetime  (void) MONO_INTERNAL;
 
 /* Stopwatch class for internal runtime use */
 typedef struct {
-	gint32 start, stop;
+	gint64 start, stop;
 } MonoStopwatch;
 
 static inline void
